Action fields should now return the context provider instead of the context
