+++ Assets/Scripts/Incidents/IIncidentActionField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 48baba9177a506049af9b6f28ce73938
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
